Updated - 20250524 https://dawnnav.pages.dev/en Increased inclusion of 5 sites.
—

People's Power (Tamil)
https://makkalathikaram.com/
All power to the working people! A revolutionary media platform in Tamil with a new democratic and socialist orientation.

Citizen News (Bengali)
https://nagorik.net/
Nagorik News is an independent progressive media outlet in Bengali.

Al-Akhbar (English Edition)
https://en.al-akhbar.com/
Al-Akhbar English is a platform for rigorous, uncompromising journalism from Palestine, Lebanon, Syria and the broader Arab region. We challenge dominant narratives that distort, erase, or silence the realities of our people and their struggles. Rooted in anti-imperialism, anti-colonialism, and popular sovereignty, our work amplifies voices from the ground and connects them to global movements across the Global South.

DSA-Liberation Faction
https://www.liberationcaucus.org/
Liberation is a Marxist-Leninist-Maoist caucus in Democratic Socialists of America.

People's Dawn (Brazil)
https://alvoradadopovo.wordpress.com/
We are a democratic-revolutionary Brazilian student movement organization. We emerged in 2018 after a group of independent students, university and secondary school activists, who did not believe in electoral horizons, sought to radicalize struggles for education and pursue a truly revolutionary path for our nation's reality. The June uprisings and hundreds of occupation strikes in secondary and tertiary education serve as a mirror for today's youth, breaking from the inertia of student representation bodies embedded in the parliamentary bureaucratic machine.