Increased inclusion of 5 sites. — Communist Party Marxist - Kenya (CPM-K) https://cpmk.org/ The Communist Party Marxist – CPM Kenya is the vanguard party of the Kenyan working class, dedicated to building an independent and sovereign socialist Kenya. The Party seeks to unite all Kenyans, regardless of ethnicity, gender, or age, to work collectively for an alternative society free from the injustices of the present capitalist system. Our vision is a socialist society that guarantees freedom, human rights, and the holistic development of all individuals.
International Anti-imperialist Bulletin
https://iabulletin.org/
On 22-23 February 2025, Internationalist Anti-War Anti-Imperialist Meeting was held in Athens. At the end of the meeting the Athens Statement was drawn up and signed by some organisations. Based on the content of this Statement, the same signatory organisations considered it appropriate and necessary to create the website International Anti-imperialist Bulletin...

Defensa del Partido Comunista del Perú
https://defensapcp.blogspot.com/

João Pedro Fragoso - Rompendo com Velhas Ideias
https://www.youtube.com/@rompendocomvelhasideias
The objective of this channel is to bring videos with analyses on economics, politics, philosophy, and current affairs from the viewpoint of the scientific ideology of the international proletariat, Marxism.

Bes D. Marx
https://www.youtube.com/@BesDMarx
Spreading Revolutionary Consciousness